Use weak functions to handle missing GBO mapping support
gbm_bo_map()/gbm_bo_unmap() are not available on libMali.
To make sure builds don't break and/or do not run properly on platforms
with Mali GPUs, check at runtime whether or not BO mapping is supported.
If it isn't, then videos won't show up on the cube (unless they are stored
in DMA-BUF buffers), and the textured cube won't work.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>