kmscube: factor out init_egl() from each cube init
The call to init_egl() can be made before calling the cube
init function. This allows it to be factored out into a common
location. struct egl is then passed in to each init() for use
as needed; the same as struct gbm.

As the cube init function no longer needs to modify struct egl
outside of setting the draw function, to keep egl const move
the draw call out of egl. It didn't make much sense in here to
begin with and seems to have been placed in this struct out
of convenience. Move the draw function to a new struct of cube
specific data.

Signed-off-by: Andrew Davis <afd@ti.com>